requirements.txt
ffmpeg-python>=0.2.0
pillow>=9.2.0
python-telegram-bot<22.0
schedule>=1.2.0
requests>=2.31.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
python-dotenv>=1.0.0
tqdm>=4.65.0
colorlog>=6.7.0
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
gTTS>=2.5.0
ollama>=0.1.6
httpx>=0.27.0,<0.28.0